Refactor argmax_program_factory.cpp - split into separate files
Split argmax_single_core and argmax_multi_core functions into separate files
for better code organization and modularity.
