+++ ttnn/cpp/ttnn/operations/reduction/CMakeLists.txt
@@ -22,6 +22,8 @@ target_sources(
     PRIVATE
         argmax/device/argmax_op.cpp
         argmax/device/argmax_program_factory.cpp
+        argmax/device/argmax_single_core_program_factory.cpp
+        argmax/device/argmax_multi_core_program_factory.cpp
         argmax/argmax.cpp
         cumprod/cumprod.cpp
         cumprod/device/cumprod_device_operation.cpp
