+++ ttnn/cpp/ttnn/operations/reduction/argmax/device/argmax_program_factory.hpp
@@ -1,20 +1,6 @@
 // SPDX-FileCopyrightText: © 2023 Tenstorrent Inc.
 //
 // SPDX-License-Identifier: Apache-2.0
-#include "ttnn/run_operation.hpp"
 
-namespace ttnn::operations::reduction::detail {
-
-using namespace tt::constants;
-
-tt::tt_metal::operation::ProgramWithCallbacks argmax_single_core(
-    const Tensor& input, const Tensor& output, std::optional<uint32_t> dim, bool keepdim);
-
-tt::tt_metal::operation::ProgramWithCallbacks argmax_multi_core(
-    const Tensor& input,
-    const Tensor& output,
-    std::optional<uint32_t> dim,
-    bool keepdim,
-    const std::optional<CoreRangeSet>& sub_core_grids);
-
-}  // namespace ttnn::operations::reduction::detail
+#include "argmax_single_core_program_factory.hpp"
+#include "argmax_multi_core_program_factory.hpp"
